+++ algopy_testing/src/algopy/arc4.py
@@ -1,18 +1,1 @@
-from __future__ import annotations
-
-import typing
-
-if typing.TYPE_CHECKING:
-    from collections.abc import Callable
-
-_P = typing.ParamSpec("_P")
-_R = typing.TypeVar("_R")
-
-
-def abimethod(
-    fn: Callable[_P, _R],
-) -> Callable[_P, _R]:
-    def wrapped(*args: _P.args, **kwargs: _P.kwargs) -> _R:
-        return fn(*args, **kwargs)
-
-    return wrapped
+from algopy_testing.arc4 import *  # noqa: F403
