+++ algopy_testing/src/algopy_testing/arc4.py
@@ -1,0 +1,22 @@
+from __future__ import annotations
+
+import typing
+
+if typing.TYPE_CHECKING:
+    from collections.abc import Callable
+
+_P = typing.ParamSpec("_P")
+_R = typing.TypeVar("_R")
+
+
+def abimethod(
+    fn: Callable[_P, _R],
+) -> Callable[_P, _R]:
+    def wrapped(*args: _P.args, **kwargs: _P.kwargs) -> _R:
+        return fn(*args, **kwargs)
+
+    return wrapped
+
+
+class ARC4Contract:
+    pass
