+++ algopy_testing/src/algopy_testing/gtxn.py
@@ -1,0 +1,6 @@
+class AssetTransferTransaction:
+    pass
+
+
+class PaymentTransaction:
+    pass
